refactor: move seasons to domain root function
